Partial fix for not sorting all the folders
The "sort all bookmarks" action works only on the "bookmarks menu".
The "Bookmarks toolbar", "Other bookmarks" and the "Mobile Bookmarks"
were not sorted (issue #12).
This commit adds "Bookmarks toolbar" and "Other bookmarks" but not
the "Mobile bookmarks". In the documentation there is not an attribute
to identify the mobile folder. Need more work on this.